Refactored streamlit
