+++ app/configs/app_config.py
@@ -1,0 +1,8 @@
+"""Config file containing static variables and prompt templates."""
+import os
+
+CONVERSATIONAL_MEMORY_SIZE = 3
+
+READABILITY_SCORE_THRESHOLD = 55.0
+
+BASE_DIR = os.path.dirname(os.path.abspath(__file__))
